swc-installation-test-2.py: Override default setup URL in JSON
So instructors can poing folks at their workshop's (or even a specific
track's) setup instructions, which are likely to be more useful than
SWC's generic instructions.  I've also updated the default URL to
point at the workshop-template source (the old URL is stale [1]).  We
could have used [2], but I'm not sure what the procedure is for
keeping that up to date, and the styling is a bit odd (e.g. row divs
that are wider than the titles they fall under).

[1]: https://github.com/wking/swc-setup-installation-test/pull/11
[2]: http://software-carpentry.org/workshops/setup.html