swc-installation-test-2.py: Prefer distro to platform.linux_distribution
platform.linux_distribution was deprecated in 2015 to be removed in
Python 3.7 [1,2].  And it may not exist at all on some Python builds
(Manjaro's [3]?).  Prefer distro (the replacement PyPI package [4,5])
on Linux if the user already has it installed.

If we can't find a distro/version, use '*' for generic help-URL
matching.  And silently skip distro/version information in
print_system_info.

[1]: http://bugs.python.org/issue1322#msg243063
[2]: https://docs.python.org/3/library/platform.html#platform.linux_distribution
[3]: https://github.com/prerit2010/swc-setup-installation-test/issues/7#issuecomment-231573710
[4]: http://bugs.python.org/issue1322#msg263896
[5]: https://pypi.python.org/pypi/distro